+++ src/js_sync.hpp
@@ -1254,7 +1254,8 @@ void SyncClass<T>::populate_sync_config(ContextType ctx, ObjectType realm_constr
         }
 #endif
 
-        ClientResyncMode clientResyncMode = realm::ClientResyncMode::Manual;
+        // default for query-based sync is manual and recover for full sync
+        ClientResyncMode clientResyncMode = (config.sync_config->is_partial) ? realm::ClientResyncMode::Manual : realm::ClientResyncMode::Recover;
         ValueType client_resync_mode_temp = Object::get_property(ctx, sync_config_object, "clientResyncMode");
         if (!Value::is_undefined(ctx, client_resync_mode_temp)) {
             std::string client_resync_mode = std::string(Value::validated_to_string(ctx, client_resync_mode_temp, "client_resync_mode"));